+++ hazel/parametric.py
@@ -106,7 +106,7 @@ class Parametric_atmosphere(General_atmosphere):
         # Check that parameters are inside borders by clipping inside the interval with a border of 1e-8
         if (self.working_mode == 'inversion'):
             for k, v in self.parameters.items():            
-                v = np.clip(v, self.ranges[k][0] + 1e-8, self.ranges[k][1] - 1e-8)
+                self.parameters[k] = np.clip(v, self.ranges[k][0] + 1e-8, self.ranges[k][1] - 1e-8)
 
     def load_reference_model(self, model_file, verbose):
         """